actualizar

+++ MedCongress/MedCongressApp/views.py
@@ -5,6 +5,7 @@ from django.views.decorators.csrf import csrf_exempt
 from collections import namedtuple
 from datetime import date,datetime
 import openpay
+from pathlib import Path
 from django.core.mail import send_mail
 import requests
 import urllib3
@@ -1911,7 +1912,9 @@ class PerfilUpdateView(FormView):
             image_result = open('MedCongressApp/static/usuarios/foto_%s_%s.png'%(nombre,nom), 'wb') # create a writable image and write the decoding result
             image_result.write(image_64_decode)
             if perfiluser.foto:
-                os.remove('MedCongressApp/static/%s'%( perfiluser.foto))
+                fileObj = Path('MedCongressApp/static/%s'%( perfiluser.foto))
+                if fileObj.is_file(): 
+                    os.remove('MedCongressApp/static/%s'%( perfiluser.foto))
             perfiluser.foto='usuarios/foto_%s_%s.png'%(nombre,nom)
         else:
             if not perfiluser.foto :
